# AngularSPA/src/app/core/api/v1/model/person.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Group } from './group';
import { Sex } from './sex';


export interface Person { 
    id?: number;
    firstName?: string | null;
    lastName?: string | null;
    birth?: string;
    sex?: Sex;
    groups?: Array<Group> | null;
}