# AngularSPA/src/app/core/api/v1/model/sex.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Sex = 1 | 2 | 3;

export const Sex = {
    NUMBER_1: 1 as Sex,
    NUMBER_2: 2 as Sex,
    NUMBER_3: 3 as Sex
};